refactor(watt): move nav-list component under shell
