+++ libs/ui-watt/src/index.ts
@@ -28,7 +28,7 @@ export * from './lib/components/button';
 export * from './lib/components/checkbox';
 export * from './lib/components/empty-state';
 export * from './lib/components/expansion';
-export * from './lib/components/nav-list';
+export * from './lib/components/shell/nav-list';
 export * from './lib/components/spinner';
 export * from './lib/components/tabs';
 export * from './lib/components/validation-message';
